handling the dots in cd command
