+++ includes/minishell.h
@@ -69,5 +69,6 @@ int pwd_function (env_list_t *env_list);
 void test_env_list (env_list_t *list);
 int cd_function (char *arg, int flag, env_list_t *env_list);
 void cd_to_home (env_list_t *env_list);
+int check_for_dots (char *arg, env_list_t *list);
 
 #endif
